1.1.0最新版 首页修改 约会修改 个人中心修改

+++ app/src/main/java/com/kuwai/ysy/module/home/bean/HomeCardBean.java
@@ -68,7 +68,7 @@ public class HomeCardBean {
         private String lastcity;
         private String constellation;
         private int weight;
-        private int distance;
+        private float distance;
         private List<AttachBean> attach;
 
         public int getUid() {
@@ -83,11 +83,11 @@ public class HomeCardBean {
             return avatar;
         }
 
-        public int getDistance() {
+        public float getDistance() {
             return distance;
         }
 
-        public void setDistance(int distance) {
+        public void setDistance(float distance) {
             this.distance = distance;
         }
